Add webpack base configuration and extend it for development and production
