+++ .eslintignore
@@ -2,6 +2,7 @@ build/
 node_modules/
 stylelintrc.config.js
 .eslintrc.js
+webpack.base.config.js
 webpack.config.js
 webpack.production.config.js
