+++ webpack.config.js
@@ -1,90 +1,18 @@
 'use strict';
 
-const path = require('path');
 const Webpack = require('webpack');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
-const stylelintConfig = require('./stylelintrc.json');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+const baseConfig = require('./webpack.base.config');
 
-const FOLDER = {
-  nodeModules: 'node_modules'
-};
+const config = Object.assign({}, baseConfig);
 
-module.exports = {
+config.entry.app.unshift('webpack-hot-middleware/client?reload=true');
 
-  entry: {
-    app: [
-      'webpack-hot-middleware/client?reload=true',
-      path.join(__dirname, 'app/js/index.jsx')
-    ],
-    vendor: ['react', 'react-dom']
-  },
+config.plugins.push(
+  new Webpack.HotModuleReplacementPlugin(),
+  new Webpack.optimize.CommonsChunkPlugin('vendor', 'js/[name].js'),
+  new Webpack.DefinePlugin({ 'process.env.NODE_ENV': JSON.stringify('development') }),
+  new ExtractTextPlugin('css/[name].css')
+);
 
-  output: {
-    path: path.join(__dirname, '/build/'),
-    publicPath: '/',
-    filename: 'js/[name].js'
-  },
-
-  devtool: 'source-map',
-
-  resolve: {
-    extensions: ['', '.js', '.jsx']
-  },
-
-  module: {
-    preLoaders: [
-      {
-        test: /\.(es6|jsx)$/,
-        loaders: ['eslint'],
-        exclude: FOLDER.nodeModules
-      }
-    ],
-
-    loaders: [
-      {
-        test: /\.(es6|jsx)$/,
-        exclude: FOLDER.nodeModules,
-        loaders: ['react-hot', 'babel']
-      },
-
-      // Build inline styles for all components
-      {
-        test: /\.scss/,
-        exclude: [FOLDER.nodeModules, path.join(__dirname, 'app/css')],
-        loader: 'style!css?modules&localIdentName=[name]__[local]!postcss!sass'
-      },
-
-      // Build main styles
-      {
-        test: /\.scss$/,
-        exclude: [FOLDER.nodeModules, path.join(__dirname, 'app/components')],
-        loader: ExtractTextPlugin.extract('style', 'css?modules&localIdentName=[name]__[local]!postcss!sass')
-      }
-    ]
-  },
-
-  postcss: function () {
-    return [
-      require('stylelint')(stylelintConfig),
-      require('autoprefixer')({browsers: ['last 4 version']})
-    ];
-  },
-
-  plugins: [
-    new CopyWebpackPlugin([
-      { from: 'app' }
-    ], {
-      ignore: ['*.html', 'css/*', 'js/*', 'components/**/**']
-    }),
-    new HtmlWebpackPlugin({
-      template: 'app/index.html',
-      inject: 'body',
-      filename: 'index.html'
-    }),
-    new Webpack.HotModuleReplacementPlugin(),
-    new Webpack.optimize.CommonsChunkPlugin('vendor', 'js/[name].js'),
-    new ExtractTextPlugin('css/[name].css')
-  ]
-};
+module.exports = config;
